requirements.txt
asgiref==3.5.2
Django==4.1.2
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-gis==1.0
install==1.3.5
Markdown==3.4.1
numpy==1.23.4
postgres==4.0
psycopg2==2.9.3
psycopg2-binary==2.9.5
psycopg2-pool==1.1
pytz==2022.6
scipy==1.9.3
smart-open==6.2.0
sqlparse==0.4.3
tzdata==2022.5
